simple cleaner and sql update

+++ IO_SubDirectories/VS_Projects_Cleaner.cs
@@ -5,44 +5,56 @@ namespace IO_SubDirectories
 {
     internal static class VS_Projects_Cleaner
     {
-        internal static void CleanAll(string solutionPath)
+        internal static void CleanAll()
         {
-            // MUST BE REFACTORED !!!
+            bool notFinished = true;
 
-            string whereToMove = "C:\\Recycle.bin";
-
-            string rootPathToSearch =
-                Environment.GetFolderPath(Environment.SpecialFolder.UserProfile) + @"\source\repos\" + solutionPath;
-
-            string[] dirNamesToFind = { "bin", "obj", "packages", ".vs" };
-
-            string[] dirNameToSkip = { ".git" };
-
-            string[] GetDirsByNames() => Directory.Exists(rootPathToSearch)
-                ? ProcessFilesDirectories.GetArrayByNames(rootPathToSearch, dirNamesToFind, dirNameToSkip)
-                : Array.Empty<string>();
-
-            //string[] mp3files = ProcessFilesDirectories.GetFilesMp3(rootPathToSearch);
-
-            foreach (var item in GetDirsByNames()) // mp3files)
+            while (notFinished)
             {
-                // TOD :
-                // REFACTORE - EXTRACT :
-                string prevDirNameOnly = Path.GetFileName(Path.GetDirectoryName(item));
-                string lastDirNameOnly = Path.GetFileName(item);
+                Console.WriteLine("Input path to project to be cleared :");
+                string rootPathToSearch = Console.ReadLine();
 
-                //string newPath = $"{whereToMove}\\{prevDirNameOnly}_{lastDirNameOnly}".Replace(' ', '_').Replace(".vs", "_vs");
-                //Console.WriteLine(newPath);
-
-                try
+                if (!Directory.Exists(rootPathToSearch))
                 {
-                    // Directory.Move(item, newPath);
-
-                    Directory.Delete(item, true);
+                    Console.WriteLine($"Not found! - {rootPathToSearch}");
+                    break;
                 }
-                catch (Exception e)
+                else
                 {
-                    Console.WriteLine(e.Message);
+                    string[] dirNamesToFind = { "bin", "obj", "packages", ".vs" };
+                    string[] dirNameToSkip = { ".git" };
+
+                    string[] GetDirsByNames() => Directory.Exists(rootPathToSearch)
+                        ? ProcessFilesDirectories.GetArrayByNames(rootPathToSearch, dirNamesToFind, dirNameToSkip)
+                        : Array.Empty<string>();
+
+                    //string[] mp3files = ProcessFilesDirectories.GetFilesMp3(rootPathToSearch);
+
+                    foreach (var item in GetDirsByNames()) // mp3files)
+                    {
+                        // TOD :
+                        // REFACTORE - EXTRACT :
+                        string prevDirNameOnly = Path.GetFileName(Path.GetDirectoryName(item));
+                        string lastDirNameOnly = Path.GetFileName(item);
+
+                        //string newPath = $"{whereToMove}\\{prevDirNameOnly}_{lastDirNameOnly}".Replace(' ', '_').Replace(".vs", "_vs");
+                        //Console.WriteLine(newPath);
+
+                        try
+                        {
+                            // Directory.Move(item, newPath);
+
+                            Console.WriteLine(item);
+
+                            Directory.Delete(item, true);
+                        }
+                        catch (Exception e)
+                        {
+                            Console.WriteLine(e.Message);
+                        }
+                    }
+
+                    notFinished = false;
                 }
             }
         }
